First implementation of Accept header parsing and prioritization
